@@ -6,15 +6,15 @@ publish_to: none
 version: 2.4.0+38
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.4 <3.0.0"
 
 dependencies:
   awesome_notifications:
     git:
       url: https://github.com/aaqibism/awesome_notifications.git
   badges: ^2.0.2
-  barcode_widget: ^2.0.2
-  barcode_image: ^2.0.1
+  barcode_widget: ^2.0.3
+  barcode_image: ^2.0.2
   cached_network_image: ^3.2.0
   carousel_slider: ^4.0.0
   charts_flutter:
@@ -23,18 +23,18 @@ dependencies:
       path: charts_flutter
   circular_profile_avatar: ^2.0.5
   clipboard: ^0.1.3
-  cloud_firestore: ^3.1.10
-  custom_refresh_indicator: ^1.0.0
+  cloud_firestore: ^4.3.1
+  custom_refresh_indicator: ^2.0.1
   dio: ^4.0.4
   dio_cache_interceptor: ^3.2.6
   dio_cache_interceptor_hive_store: ^3.1.1
   dotted_border: ^2.0.0+2
   extended_image: ^6.0.2+1
-  firebase_analytics: ^9.1.2
-  firebase_core: ^1.13.1
-  firebase_in_app_messaging: ^0.6.0+9
-  firebase_messaging: ^11.2.11
-  firebase_remote_config: ^2.0.2
+  firebase_analytics: ^10.1.0
+  firebase_core: ^2.4.1
+  firebase_in_app_messaging: ^0.7.0+10
+  firebase_messaging: ^14.2.1
+  firebase_remote_config: ^3.0.9
   flare_flutter: ^3.0.2
   flutter:
     sdk: flutter
@@ -44,7 +44,7 @@ dependencies:
   flutter_masked_text:
     git:
       url: https://github.com/vitorhm/flutter-masked-text.git
-  flutter_secure_storage: ^5.0.2
+  flutter_secure_storage: ^7.0.1
   flutter_spinkit: ^5.1.0
   flutter_windowmanager: ^0.2.0
   font_awesome_flutter: ^10.0.0
@@ -55,25 +55,25 @@ dependencies:
   image_picker: ^0.8.4+2
   in_app_review: ^2.0.3
   mdi: ^5.0.0-nullsafety.0
-  package_info: ^2.0.2
+  package_info_plus: ^3.0.2
   path_provider: ^2.0.2
-  permission_handler: ^9.2.0
-  photo_view: ^0.13.0
+  permission_handler: ^10.2.0
+  photo_view: ^0.14.0
   qr_flutter: ^4.0.0
   recase: ^4.0.0
-  responsive_framework: ^0.1.0
+  responsive_framework: ^0.2.0
   screenshot: ^1.2.3
   sentry: ^6.0.0
-  share_plus: ^4.0.1
+  share_plus: ^6.3.0
   shared_preferences: ^2.0.7
   simple_gesture_detector: ^0.2.0
-  url_launcher: ^6.1.2
+  url_launcher: ^6.1.7
   video_player: ^2.1.0
   dio_http_cache: ^0.3.0
-  image: ^3.1.3
+  image: ^3.1.0
   intl: ^0.17.0
   get_storage: ^2.0.3
-  hexcolor: ^2.0.7
+  hexcolor: ^3.0.1
   
 
 dependency_overrides:
@@ -83,7 +83,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   import_sorter: ^4.6.0
-  lint: ^1.5.3
+  lint: ^2.0.1
 
 flutter:
   uses-material-design: true
