@@ -1,22 +1,13 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 
-import 'package:get/get.dart';
-
-import 'package:mi_utem/themes/theme.dart';
-import 'package:mi_utem/widgets/acerca_aplicacion_content.dart';
-import 'package:mi_utem/widgets/acerca_screen.dart';
-import 'package:mi_utem/widgets/loading_indicator.dart';
-
 class CustomAlertDialog extends StatefulWidget {
-  String titulo;
-  String? emoji;
-  String descripcion;
-  String confirmarTextoBoton;
-  String cancelarTextoBoton;
-  VoidCallback? onCancelar;
-  VoidCallback? onConfirmar;
+  final String titulo;
+  final String? emoji;
+  final String descripcion;
+  final String confirmarTextoBoton;
+  final String cancelarTextoBoton;
+  final VoidCallback? onCancelar;
+  final VoidCallback? onConfirmar;
 
   CustomAlertDialog(
       {Key? key,
@@ -80,13 +71,11 @@ class _CustomAlertDialogState extends State<CustomAlertDialog> {
                                   fontSize: 50,
                                 ),
                               ),
-                            if (widget.descripcion != null)
-                              Container(height: 20),
-                            if (widget.descripcion != null)
-                              Text(
-                                widget.descripcion,
-                                textAlign: TextAlign.center,
-                              ),
+                            Container(height: 20),
+                            Text(
+                              widget.descripcion,
+                              textAlign: TextAlign.center,
+                            ),
                             Container(height: 20),
                             Wrap(
                               spacing: 10,
